+++ app/src/main/java/myapplication/second/workinghourmanagement/member/ResetPasswdActivity.kt
@@ -1,18 +1,18 @@
-package myapplication.second.workinghourmanagement
+package myapplication.second.workinghourmanagement.member
 
 import android.os.Bundle
 import androidx.appcompat.app.ActionBar
 import androidx.appcompat.app.AppCompatActivity
 import android.view.MenuItem
 import androidx.databinding.DataBindingUtil
-import myapplication.second.workinghourmanagement.databinding.ActivityStaffJoinBinding
+import myapplication.second.workinghourmanagement.R
+import myapplication.second.workinghourmanagement.databinding.ActivityCommonResetpwBinding
 
-class StaffJoinActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityStaffJoinBinding
-
+class ResetPasswdActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityCommonResetpwBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_staff_join)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_common_resetpw)
 
         setSupportActionBar(binding.toolbar)
         val actionBar: ActionBar = supportActionBar!!
